ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
python-multipart==0.0.6
requests==2.31.0
